+++ app/frontend/src/dashboard/dashboard.template.html
@@ -1,9 +1,11 @@
-<div ng-if="$ctrl.user.isStudent">
-    <student-dashboard></student-dashboard>
-</div>
-<div ng-if="$ctrl.user.isTeacher">
-    <teacher-dashboard></teacher-dashboard>
-</div>
-<div ng-if="$ctrl.user.isAdmin">
-    <reservations user-role="admin"></reservations>
+<div style="min-height: 700px">
+    <div ng-if="$ctrl.user.isStudent">
+        <student-dashboard></student-dashboard>
+    </div>
+    <div ng-if="$ctrl.user.isTeacher">
+        <teacher-dashboard></teacher-dashboard>
+    </div>
+    <div ng-if="$ctrl.user.isAdmin">
+        <admin-reservations user-role="admin"></admin-reservations>
+    </div>
 </div>
