+++ app/frontend/src/reservation/admin/adminReservations.template.html
@@ -23,47 +23,41 @@
                 <div class="row top-margin-1">
                     <div class="col-md-4">
                         <strong>{{'sitnet_exam_owner' | translate}}</strong>
-                        <input class="wide-select" ui-select2="$ctrl.teacherOptions"
-                                ng-model="$ctrl.selection.ownerId" ng-click="$ctrl.query()">
-                        </input>
+                        <drop-down-select options="$ctrl.teacherOptions" on-select="$ctrl.ownerChanged(value)">
+                        </drop-down-select>
                     </div>
                     <div class="col-md-4">
                         <strong>{{'sitnet_exam_state' | translate}}</strong>
-                        <input class="wide-select" ui-select2="$ctrl.stateOptions"
-                                ng-model="$ctrl.selection.state" ng-click="$ctrl.query()">
-                        </input>
+                        <drop-down-select options="$ctrl.stateOptions" on-select="$ctrl.stateChanged(value)">
+                        </drop-down-select>
                     </div>
                     <div class="col-md-4">
                         <strong>{{'sitnet_student' | translate}}</strong>
-                        <input class="wide-select" ui-select2="$ctrl.studentOptions"
-                                ng-model="$ctrl.selection.studentId" ng-click="$ctrl.query()">
-                        </input>
+                        <drop-down-select options="$ctrl.studentOptions" on-select="$ctrl.studentChanged(value)">
+                        </drop-down-select>
                     </div>
                 </div>
                 <div class="row top-margin-1">
                     <div class="col-md-4">
                         <strong>{{'sitnet_room' | translate}}</strong>
-                        <input class="wide-select" ui-select2="$ctrl.roomOptions"
-                                ng-model="$ctrl.selection.roomId" ng-click="$ctrl.roomChanged()">
-                        </input>
+                        <drop-down-select options="$ctrl.roomOptions" on-select="$ctrl.roomChanged(value)">
+                        </drop-down-select>
                     </div>
                     <div class="col-md-4">
                         <strong>{{'sitnet_exam_machine' | translate}}</strong>
-                        <input class="wide-select" ui-select2="$ctrl.machineOptions"
-                                ng-model="$ctrl.selection.machineId" ng-click="$ctrl.query()">
-                        </input>
+                        <drop-down-select options="$ctrl.machineOptions" on-select="$ctrl.machineChanged(value)">
+                        </drop-down-select>
                     </div>
                     <div class="col-md-4">
                         <strong>{{'sitnet_exam' | translate}}</strong>
-                        <input class="wide-select" ui-select2="$ctrl.examOptions"
-                                ng-model="$ctrl.selection.examId" ng-click="$ctrl.query()">
-                        </input>
+                        <drop-down-select options="$ctrl.examOptions" on-select="$ctrl.examChanged(value)">
+                        </drop-down-select>
                     </div>
                 </div>
             </form>
         </div>
     </div>
 
-    <div class="top-margin-1" ng-init="$ctrl.query()" ng-include="$ctrl.reservationDetails"></div>
+    <reservation-detail ng-if="$ctrl.reservations" reservations="$ctrl.reservations" is-admin-view="true"></reservation-detail>
 
 </div>
