+++ app/frontend/src/reservation/admin/removeReservationDialog.template.html
@@ -4,17 +4,17 @@
     </div>
     <div class="modal-body">
         <strong>{{'sitnet_message' | translate}}</strong>
-        <textarea class="form-control" ng-model="message.text"
+        <textarea class="form-control" ng-model="$ctrl.message.text"
                   rows="3">
         </textarea>
     </div>
     <div class="modal-footer">
         <div class="col-md-6">
-            <button class="btn btn-sm btn-danger pull-left" ng-click="cancel()">{{'sitnet_button_cancel' | translate}}
+            <button class="btn btn-sm btn-danger pull-left" ng-click="$ctrl.cancel()">{{'sitnet_button_cancel' | translate}}
             </button>
         </div>
         <div class="col-md-6">
-            <button class="btn btn-sm btn-primary" ng-click="ok()">
+            <button class="btn btn-sm btn-primary" ng-click="$ctrl.ok()">
                 {{'sitnet_send' | translate}}
             </button>
         </div>
