+++ app/frontend/src/utility/select/dropDownSelect.template.html
@@ -1,0 +1,30 @@
+<div class="dropdown" uib-dropdown>
+    <button uib-dropdown-toggle class="btn btn-default btn-block dropdown-toggle" type="button"
+            data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
+        <span style="float: left">{{$ctrl.selected.label || $ctrl.placeholder | translate}}</span>
+        <div style="float: right">
+            <span class="caret"></span>
+        </div>
+    </button>
+    <ul class="dropdown-menu scrollable-menu" style="padding-left: 0; min-width: 17em" role="menu">
+        <li role="presentation" class="input-group">
+            <input ng-model="$ctrl.searchFilter" class="form-control"
+                   ng-click="$event.stopPropagation()"
+                   placeholder="{{'sitnet_search' | translate }}">
+
+            <div class="input-group-addon">
+                <i class="fa fa-search fa-fw"></i>
+            </div>
+        </li>
+        <li role="presentation">
+            <a class="menu-item" ng-click="$ctrl.clearSelection()">{{$ctrl.placeholder | translate}}</a>
+        </li>
+        <li ng-repeat="opt in $ctrl.options | filter: $ctrl.optionFilter" role="presentation"
+            ng-class="$ctrl.getClasses(opt)">
+            <a ng-if="!opt.isHeader" role="menuitem" ng-click="$ctrl.selectOption(opt)">
+                {{opt.label | translate | truncate: 40}}
+            </a>
+            <span ng-if="opt.isHeader">{{opt.label}}</span>
+        </li>
+    </ul>
+</div>
